chore: Refactor models to include into migrations
